rooms, participants and messages

+++ pr/base/views.py
@@ -1,20 +1,93 @@
 from django.shortcuts import render, redirect
-from .models import Room
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+#from django.contrib.messages import constants as messages
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic, Message
 from .forms import RoomForm
 #from django.http import HttpResponse, HttpResponseNotFound
 #from . import views
 
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password does not exist')
+    context = {'page':page}
+    return render(request, 'base/login_register.html', context)
+
+def logoutPage(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    page = 'register'
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+    return render(request, 'base/login_register.html', {'form':form, 'page':page})
+
 def home(request):
-    rooms = Room.objects.all()
+    q = request.GET.get('q')
+    if q is None:
+        q = ''
+    if q == '':
+        rooms = Room.objects.all()
+    else:
+        rooms = Room.objects.filter(
+            Q(topic__name__icontains=q) |
+            Q(name__icontains=q) |
+            Q(description__icontains=q)
+        )
+        #rooms = Room.objects.filter(topic__name__icontains=q)
+        #rooms = Room.objects.filter(topic__name=q)
+    topics = Topic.objects.all()
+    room_count = rooms.count()
     #return HttpResponse("Hello, world. You're at the pr index.")
-    return render(request, 'base/home.html', {'rooms' : rooms})
+    context = {'rooms' : rooms, 'topics': topics, 'room_count': room_count}
+    return render(request, 'base/home.html', context)
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room' : room}
+    messages = room.message_set.all().order_by('-created_at')
+    participants = room.participants.all()
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user = request.user,
+            room = room,
+            body = request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+    context = {'room' : room, 'room_messages': messages, 'participants': participants}
     return render(request, 'base/room.html', context)
 
-def create_room(request):
+@login_required(login_url='login_page')
+def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
         form = RoomForm(request.POST)
@@ -25,9 +98,12 @@ def create_room(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-def update_room(request, pk):
+@login_required(login_url='login_page')
+def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -36,12 +112,27 @@ def update_room(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-def delete_room(request, pk):
+@login_required(login_url='login_page')
+def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     context = {'obj': room}
+    return render(request, 'base/delete.html', context)
+
+
+@login_required(login_url='login_page')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here!')
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    context = {'obj': message}
     return render(request, 'base/delete.html', context)
 
 def custom_404_view(request, exception):
